requirements.txt
ultralytics==8.3.0
torch>=2.2
torchvision>=0.17
pycocotools
opencv-python
numpy
tqdm
matplotlib
PyYAML
